+++ pkg/providers/appsec/appsec_test.go
@@ -913,6 +913,26 @@ func (p *mockappsec) GetPenaltyBoxes(ctx context.Context, params appsec.GetPenal
 	return args.Get(0).(*appsec.GetPenaltyBoxesResponse), args.Error(1)
 }
 
+func (p *mockappsec) GetEvalPenaltyBox(ctx context.Context, params appsec.GetPenaltyBoxRequest) (*appsec.GetPenaltyBoxResponse, error) {
+	args := p.Called(ctx, params)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*appsec.GetPenaltyBoxResponse), args.Error(1)
+}
+
+func (p *mockappsec) UpdateEvalPenaltyBox(ctx context.Context, params appsec.UpdatePenaltyBoxRequest) (*appsec.UpdatePenaltyBoxResponse, error) {
+	args := p.Called(ctx, params)
+
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+
+	return args.Get(0).(*appsec.UpdatePenaltyBoxResponse), args.Error(1)
+}
+
 func (p *mockappsec) GetSecurityPolicies(ctx context.Context, params appsec.GetSecurityPoliciesRequest) (*appsec.GetSecurityPoliciesResponse, error) {
 	args := p.Called(ctx, params)