+++ pkg/providers/appsec/data_akamai_appsec_eval_groups_test.go
@@ -56,7 +56,7 @@ func TestAccAkamaiEvalGroups_data_error_retrieving_eval_groups(t *testing.T) {
 		json.Unmarshal([]byte(loadFixtureBytes("testdata/TestResConfiguration/LatestConfiguration.json")), &configs)
 
 		client.On("GetEvalGroups",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			appsec.GetAttackGroupsRequest{ConfigID: 43253, Version: 7, PolicyID: "AAAA_81230", Group: "SQL"},
 		).Return(nil, fmt.Errorf("GetEvalGroups failed"))