+++ pkg/providers/appsec/resource_akamai_appsec_eval_penalty_box.go
@@ -2,7 +2,6 @@ package appsec
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"strconv"
 
@@ -18,12 +17,12 @@ import (
 // appsec v1
 //
 // https://developer.akamai.com/api/cloud_security/application_security/v1.html
-func resourcePenaltyBox() *schema.Resource {
+func resourceEvalPenaltyBox() *schema.Resource {
 	return &schema.Resource{
-		CreateContext: resourcePenaltyBoxCreate,
-		ReadContext:   resourcePenaltyBoxRead,
-		UpdateContext: resourcePenaltyBoxUpdate,
-		DeleteContext: resourcePenaltyBoxDelete,
+		CreateContext: resourceEvalPenaltyBoxCreate,
+		ReadContext:   resourceEvalPenaltyBoxRead,
+		UpdateContext: resourceEvalPenaltyBoxUpdate,
+		DeleteContext: resourceEvalPenaltyBoxDelete,
 		CustomizeDiff: customdiff.All(
 			VerifyIDUnchanged,
 		),
@@ -32,20 +31,24 @@ func resourcePenaltyBox() *schema.Resource {
 		},
 		Schema: map[string]*schema.Schema{
 			"config_id": {
-				Type:     schema.TypeInt,
-				Required: true,
+				Type:        schema.TypeInt,
+				Required:    true,
+				Description: "Unique identifier of the security configuration.",
 			},
 			"security_policy_id": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Unique identifier of the security policy.",
 			},
 			"penalty_box_protection": {
-				Type:     schema.TypeBool,
-				Required: true,
+				Type:        schema.TypeBool,
+				Required:    true,
+				Description: "Enables or disables penalty box.",
 			},
 			"penalty_box_action": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Penalty box action applied to requests from clients in the penalty box.",
 				ValidateDiagFunc: validation.ToDiagFunc(validation.StringInSlice([]string{
 					string(appsec.ActionTypeDeny),
 					string(appsec.ActionTypeAlert),
@@ -56,17 +59,17 @@ func resourcePenaltyBox() *schema.Resource {
 	}
 }
 
-func resourcePenaltyBoxCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceEvalPenaltyBoxCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
 	client := inst.Client(meta)
-	logger := meta.Log("APPSEC", "resourcePenaltyBoxCreate")
-	logger.Debugf("in resourcePenaltyBoxCreate")
+	logger := meta.Log("APPSEC", "resourceEvalPenaltyBoxCreate")
+	logger.Debugf("in resourceEvalPenaltyBoxCreate")
 
 	configID, err := tools.GetIntValue("config_id", d)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	version, err := getModifiableConfigVersion(ctx, configID, "penaltyBoxAction", m)
+	version, err := getModifiableConfigVersion(ctx, configID, "evalPenaltyBox", m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -75,11 +78,11 @@ func resourcePenaltyBoxCreate(ctx context.Context, d *schema.ResourceData, m int
 		return diag.FromErr(err)
 	}
 	penaltyboxprotection, err := tools.GetBoolValue("penalty_box_protection", d)
-	if err != nil && !errors.Is(err, tools.ErrNotFound) {
+	if err != nil {
 		return diag.FromErr(err)
 	}
 	penaltyboxaction, err := tools.GetStringValue("penalty_box_action", d)
-	if err != nil && !errors.Is(err, tools.ErrNotFound) {
+	if err != nil {
 		return diag.FromErr(err)
 	}
 
@@ -91,22 +94,22 @@ func resourcePenaltyBoxCreate(ctx context.Context, d *schema.ResourceData, m int
 		Action:               penaltyboxaction,
 	}
 
-	_, err = client.UpdatePenaltyBox(ctx, createPenaltyBox)
+	_, err = client.UpdateEvalPenaltyBox(ctx, createPenaltyBox)
 	if err != nil {
-		logger.Errorf("calling 'createPenaltyBox': %s", err.Error())
+		logger.Errorf("calling 'createEvalPenaltyBox': %s", err.Error())
 		return diag.FromErr(err)
 	}
 
 	d.SetId(fmt.Sprintf("%d:%s", createPenaltyBox.ConfigID, createPenaltyBox.PolicyID))
 
-	return resourcePenaltyBoxRead(ctx, d, m)
+	return resourceEvalPenaltyBoxRead(ctx, d, m)
 }
 
-func resourcePenaltyBoxRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceEvalPenaltyBoxRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
 	client := inst.Client(meta)
-	logger := meta.Log("APPSEC", "resourcePenaltyBoxRead")
-	logger.Debugf("in resourcePenaltyBoxRead")
+	logger := meta.Log("APPSEC", "resourceEvalPenaltyBoxRead")
+	logger.Debugf("in resourceEvalPenaltyBoxRead")
 
 	iDParts, err := splitID(d.Id(), 2, "configID:securityPolicyID")
 	if err != nil {
@@ -128,7 +131,7 @@ func resourcePenaltyBoxRead(ctx context.Context, d *schema.ResourceData, m inter
 		PolicyID: policyID,
 	}
 
-	penaltybox, err := client.GetPenaltyBox(ctx, getPenaltyBox)
+	penaltybox, err := client.GetEvalPenaltyBox(ctx, getPenaltyBox)
 	if err != nil {
 		logger.Errorf("calling 'getPenaltyBox': %s", err.Error())
 		return diag.FromErr(err)
@@ -150,11 +153,11 @@ func resourcePenaltyBoxRead(ctx context.Context, d *schema.ResourceData, m inter
 	return nil
 }
 
-func resourcePenaltyBoxUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceEvalPenaltyBoxUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
 	client := inst.Client(meta)
-	logger := meta.Log("APPSEC", "resourcePenaltyBoxUpdate")
-	logger.Debugf("in resourcePenaltyBoxUpdate")
+	logger := meta.Log("APPSEC", "resourceEvalPenaltyBoxUpdate")
+	logger.Debugf("in resourceEvalPenaltyBoxUpdate")
 
 	iDParts, err := splitID(d.Id(), 2, "configID:securityPolicyID")
 	if err != nil {
@@ -164,17 +167,17 @@ func resourcePenaltyBoxUpdate(ctx context.Context, d *schema.ResourceData, m int
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	version, err := getModifiableConfigVersion(ctx, configID, "penaltyBoxAction", m)
+	version, err := getModifiableConfigVersion(ctx, configID, "evalPenaltyBox", m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 	policyID := iDParts[1]
 	penaltyboxprotection, err := tools.GetBoolValue("penalty_box_protection", d)
-	if err != nil && !errors.Is(err, tools.ErrNotFound) {
+	if err != nil {
 		return diag.FromErr(err)
 	}
 	penaltyboxaction, err := tools.GetStringValue("penalty_box_action", d)
-	if err != nil && !errors.Is(err, tools.ErrNotFound) {
+	if err != nil {
 		return diag.FromErr(err)
 	}
 
@@ -186,20 +189,20 @@ func resourcePenaltyBoxUpdate(ctx context.Context, d *schema.ResourceData, m int
 		Action:               penaltyboxaction,
 	}
 
-	_, err = client.UpdatePenaltyBox(ctx, updatePenaltyBox)
+	_, err = client.UpdateEvalPenaltyBox(ctx, updatePenaltyBox)
 	if err != nil {
-		logger.Errorf("calling 'updatePenaltyBox': %s", err.Error())
+		logger.Errorf("calling 'updateEvalPenaltyBox': %s", err.Error())
 		return diag.FromErr(err)
 	}
 
-	return resourcePenaltyBoxRead(ctx, d, m)
+	return resourceEvalPenaltyBoxRead(ctx, d, m)
 }
 
-func resourcePenaltyBoxDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceEvalPenaltyBoxDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
 	client := inst.Client(meta)
-	logger := meta.Log("APPSEC", "resourcePenaltyBoxDelete")
-	logger.Debugf("in resourcePenaltyBoxDelete")
+	logger := meta.Log("APPSEC", "resourceEvalPenaltyBoxDelete")
+	logger.Debugf("in resourceEvalPenaltyBoxDelete")
 
 	iDParts, err := splitID(d.Id(), 2, "configID:securityPolicyID")
 	if err != nil {
@@ -209,7 +212,7 @@ func resourcePenaltyBoxDelete(ctx context.Context, d *schema.ResourceData, m int
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	version, err := getModifiableConfigVersion(ctx, configID, "penaltyBoxAction", m)
+	version, err := getModifiableConfigVersion(ctx, configID, "evalPenaltyBox", m)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -223,9 +226,9 @@ func resourcePenaltyBoxDelete(ctx context.Context, d *schema.ResourceData, m int
 		Action:               string(appsec.ActionTypeNone),
 	}
 
-	_, err = client.UpdatePenaltyBox(ctx, removePenaltyBox)
+	_, err = client.UpdateEvalPenaltyBox(ctx, removePenaltyBox)
 	if err != nil {
-		logger.Errorf("calling 'removePenaltyBox': %s", err.Error())
+		logger.Errorf("calling 'removeEvalPenaltyBox': %s", err.Error())
 		return diag.FromErr(err)
 	}
 	d.SetId("")