+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## 2.x.x (June XX, 2022)
 
+#### FEATURES/ENHANCEMENTS:
+* APPSEC
+  * Added penalty box support for security policy in evaluation mode 
+
 #### BUG FIXES:
 
 * APPSEC
   * Fix drift in `EffectiveTimePeriod`, `SamplingRate`, `LoggingOptions`, and `Operation` fields of custom rule resource.
   * Fix crash when eval rule API returns an error.
+
 
 ## 2.1.1 (Jun 9, 2022)